app/requirements.txt
python-dotenv
streamlit
streamlit-authenticator
streamlit-chat
langchain
openai
tiktoken
markdownify
faiss-cpu